implement api endpoint for statistics and correct statistics updating

+++ websrv-flask/app/model/QuestionStatistic.py
@@ -28,13 +28,13 @@ class QuestionStatistic(PersistentObject):
             self.q3 = self.median(result_list[int(result_list_length / 2):])
 
         if result_list_length % 4 == 1:
-            n = (result_list_length - 1) / 4
-            self.q1 = 0.25 * result_list[n - 1].answer_value + 0.75 * result_list[n].answer_value
-            self.q3 = 0.75 * result_list[3*n].answer_value + 0.75 * result_list[3*n + 1].answer_value
+            n = int((result_list_length - 1) / 4)
+            self.q1 = 0.25 * int(result_list[n - 1].answer_value) + 0.75 * int(result_list[n].answer_value)
+            self.q3 = 0.75 * int(result_list[3*n].answer_value) + 0.75 * int(result_list[3*n + 1].answer_value)
         if result_list_length % 4 == 3:
-            n = (result_list_length - 3) / 4
-            self.q1 = 0.25 * result_list[n].answer_value + 0.75 * result_list[n + 1].answer_value
-            self.q3 = 0.75 * result_list[3*n + 1].answer_value + 0.75 * result_list[3*n + 2].answer_value
+            n = int((result_list_length - 3) / 4)
+            self.q1 = 0.25 * int(result_list[n].answer_value) + 0.75 * int(result_list[n + 1].answer_value)
+            self.q3 = 0.75 * int(result_list[3*n + 1].answer_value) + 0.75 * int(result_list[3*n + 2].answer_value)
 
     def median(self, values):
         """
